requirements.txt
openai
fastapi
uvicorn
python-dotenv==0.19.2
